// app/(dashboard)/certificates/eicr/[id]/update/schedule-of-inspections/distribution-equipment/schema.ts
import { z } from "zod";

export const UpdateDistributionEquipmentSchema = z.object({
  item_4_1: z.string(),
  item_4_2: z.string(),
  item_4_3: z.string(),
  item_4_4: z.string(),
  item_4_5: z.string(),
  item_4_6: z.string(),
  item_4_7: z.string(),
  item_4_8: z.string(),
  item_4_9: z.string(),
  item_4_10: z.string(),
  item_4_11: z.string(),
  item_4_12: z.string(),
  item_4_13: z.string(),
  item_4_14: z.string(),
  item_4_15: z.string(),
  item_4_16: z.string(),
  item_4_17: z.string(),
  item_4_18: z.string(),
  item_4_19: z.string(),
  item_4_20: z.string(),
  item_4_21: z.string(),
  item_4_22: z.string(),
  item_4_23: z.string(),
  item_4_24: z.string(),
  item_4_25: z.string(),
});